requirements.txt
python-telegram-bot==20.6
Flask==2.3.2
Flask-Cors==3.0.10
apscheduler==3.10.1
